# src/main/java/com/project/ecommerce/perfume/dto/order/OrderItemResponse.java
package com.project.ecommerce.perfume.dto.order;

import com.project.ecommerce.perfume.dto.perfume.PerfumeResponse;
import lombok.Getter;
import lombok.Setter;

@Getter
@Setter
public class OrderItemResponse {
    private Long id;
    private Long amount;
    private Long quantity;
    private PerfumeResponse perfume;
}